requirements.txt
streamlit
pandas
numpy
matplotlib
pydantic
python-dotenv
requests
openai>=1.40.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
groq>=0.9.0


